+++ EKIHIRE/EkiHire.Core/Domain/Entities/UserRole.cs
@@ -2,7 +2,7 @@
 
 namespace EkiHire.Core.Domain.Entities
 {
-    public class UserRole : IdentityUserRole<int>
+    public class UserRole : IdentityUserRole<long>
     {
         public virtual User User { get; set; }
         public virtual Role Role { get; set; }
